% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_line_tweets.R
\name{extract_entities_dt}
\alias{extract_entities_dt}
\title{Extract Entities Data.Table from Twitter Data}
\usage{
extract_entities_dt(tweets)
}
\arguments{
\item{tweets}{A data.table containing Twitter data, including tweet_id or user_id and entities columns.}
}
\value{
A data.table containing entities information extracted from Twitter data in wide format.
}
\description{
This function extracts entities information from Twitter data and organizes it into a data.table in wide format.
}
\details{
This function drops empty rows, subsets the data.table to tweet_id and entities columns, unnests the entities column into a longer format,
and pivots the data into wide format. It creates a unique row identifier for each tweet_id - entities_id pair to ensure uniqueness in the resulting data.table.
}
