% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_line_tweets.R
\name{extract_hashtags_dt}
\alias{extract_hashtags_dt}
\title{Extract Hashtags Data.Table from Entities Data.Table}
\usage{
extract_hashtags_dt(entities, api_version = "v2")
}
\arguments{
\item{entities}{A data.table containing entities information extracted from Twitter data, typically obtained using the extract_entities_dt function.}

\item{api_version}{A character vector specifying the API version. Default is v2.}
}
\value{
A data.table containing hashtags extracted from the tweets, along with related information such as tweet_id, row_id, tag, start, and end.
}
\description{
This function extracts hashtags from tweets' entities data.table and organizes them into a separate data.table.
}
\details{
This function evaluates if there are tweets with several hashtags, resulting in a hashtags.tag column. It then subsets the data into tweets
containing one hashtag and tweets with several hashtags. Column names starting with "hashtags" are harmonized, and the prefix "hashtags" is removed.
Empty observations are dropped, and the resulting data.table is indexed by tweet_id and tag for efficient querying.
}
