% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_line_tweets.R
\name{extract_mentions_dt}
\alias{extract_mentions_dt}
\title{Extract Mentions Data.Table from Entities Data.Table}
\usage{
extract_mentions_dt(entities, api_version = "v2")
}
\arguments{
\item{entities}{A data.table containing entities information extracted from Twitter data, typically obtained using the extract_entities_dt function.}

\item{api_version}{A character vector specifying the API version. Default is v2.}
}
\value{
A data.table containing mentions extracted from the tweets, along with related information such as tweet_id, row_id, username, id, start, and end.
}
\description{
This function extracts mentions from tweets' entities data.table and organizes them into a separate data.table.
}
\details{
This function evaluates if there are tweets with several mentions, resulting in a mentions.id column. It then subsets the data into tweets containing
one mention and tweets with several mentions. Column names starting with "mentions" are harmonized, and the prefix "mentions." is removed. Empty observations are dropped,
and the resulting data.table is indexed by tweet_id, username, and id for efficient querying.
}
