% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_line_tweets.R
\name{extract_URL_dt}
\alias{extract_URL_dt}
\title{Extract URL Data.Table from Twitter Data}
\usage{
extract_URL_dt(tweets)
}
\arguments{
\item{tweets}{A data.table containing Twitter data, including tweet_id and entities columns.}
}
\value{
A data.table containing URL information extracted from Twitter data.
}
\description{
This function extracts URL information from Twitter data and organizes it into a data.table format.
}
\details{
This function drops empty rows, subsets the data.table to tweet_id and entities columns, unnests the entities column into a longer format, and pivots the data into wide format. If multiple URLs are present in a tweet, it separates them into individual rows and harmonizes the column names. It also drops empty observations and creates a new column "clear_url" based on the condition provided. Additionally, it extracts domain names from the expanded_url column.
}
