% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_preprocessed.R
\name{filter_batchwise}
\alias{filter_batchwise}
\title{Load, Filter, and Save Tweets Data batchwise}
\usage{
filter_batchwise(
  file_paths,
  save_path,
  drop_type = c("retweeted", "quoted"),
  filter_all_dts = T,
  return = T
)
}
\arguments{
\item{file_paths}{A character vector of file paths to the .rds files containing the data.tables to be processed.}

\item{save_path}{A character string specifying the directory path where the filtered .rds files will be saved.}

\item{drop_type}{A character vector indicating the types of tweets to filter out. Default is \code{c("retweeted", "quoted")}.}

\item{filter_all_dts}{Logical value indicating whether to filter all data.tables in the list based on the remaining 'tweet_id's in 'tweets'. Default is \code{TRUE}.}

\item{return}{Logical value indicating whether to return the filtered data.tables as a list. Default is \code{TRUE}.}
}
\value{
If \code{return = TRUE}, a list of filtered data.tables is returned. If \code{return = FALSE}, the function only saves the filtered data to the specified \code{save_path}.
}
\description{
This function loads a list of data files, filters the 'tweets' data.table in each file based on the specified types, and saves the filtered data back to the disk. The function can also filter other data.tables in the list based on the remaining 'tweet_id's in 'tweets'.
}
