% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_preprocessed.R
\name{filter_tweets}
\alias{filter_tweets}
\title{Filter Tweets Based on Referenced Types}
\usage{
filter_tweets(
  preprocessed,
  drop_type = c("retweeted", "quoted"),
  filter_all_dts = TRUE
)
}
\arguments{
\item{preprocessed}{A list containing two data.tables: \code{tweets} and \code{referenced}. The \code{tweets} data.table should contain the data to be filtered. The \code{referenced} data.table should contain a \code{tweet_id} column and a \code{type} column indicating the type of reference.}

\item{drop_type}{A character vector specifying the types of references to drop from the \code{tweets} data.table. Default is \code{c("retweeted", "quoted")}.}

\item{filter_all_dts}{Logical. Should the other data.tables \code{referenced}, \code{urls}, \code{mentions}, and \code{hashtags} be filtered by remaining \code{tweet_id}?}
}
\value{
A list containing the modified \code{tweets} data.table with rows removed based on the specified \code{drop_type}. The \code{referenced} data.table remains unchanged.
}
\description{
This function filters out observations from the \code{tweets} data.table in a list object based on the type of references specified. It drops rows in \code{tweets} where the \code{tweet_id} matches those in the \code{referenced} data.table with the specified types (e.g., "retweeted" or "quoted").
}
