% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_batches.R
\name{load_preprocess_users_jsonl_multi}
\alias{load_preprocess_users_jsonl_multi}
\title{This function loads user data from JSONL files and preprocesses them.}
\usage{
load_preprocess_users_jsonl_multi(
  data_dir = NULL,
  file_paths = NULL,
  save_dir = NULL,
  bind = TRUE,
  api_version = "v2",
  ...
)
}
\arguments{
\item{data_dir}{A character string specifying the directory containing JSONL files. Defaults to \code{NULL}.}

\item{file_paths}{A character vector specifying the paths of JSONL files. Defaults to \code{NULL}.}

\item{save_dir}{A character string specifying the directory to save preprocessed data. Defaults to \code{NULL}.}

\item{bind}{Logical. Should results be bound to one list of data tables or returned as list of lists? Default is TRUE.}

\item{api_version}{A character vector specifying the API version. Default is v2.}

\item{...}{Additional arguments to be passed to \code{\link[jsonlite]{fromJSON}} when loading JSONL files.}
}
\value{
A list containing preprocessed tweet data.
}
\description{
This function loads user data from JSONL files and preprocesses them.
}
