% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_by_line.R
\name{load_tweets_multi_json_line}
\alias{load_tweets_multi_json_line}
\title{Load Tweets from Multiple JSON Files}
\usage{
load_tweets_multi_json_line(data_dir, ...)
}
\arguments{
\item{data_dir}{Directory containing the JSON files.}

\item{...}{Additional arguments to pass to \code{\link{load_tweets_json_line}} function.}
}
\value{
A data.table containing tweets from all JSON files.
}
\description{
Load tweets from multiple JSON files located in the specified directory.
}
