% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_by_line.R
\name{load_users_jsonl}
\alias{load_users_jsonl}
\title{Load Users from JSONL File}
\usage{
load_users_jsonl(
  file_path,
  num_threads = NULL,
  n_max = Inf,
  skip_empty_rows = TRUE,
  query = NULL,
  query_error_ok = TRUE,
  on_query_error = NULL,
  parse_error_ok = TRUE,
  on_parse_error = NULL,
  empty_array = NULL,
  empty_object = NULL,
  max_simplify_lvl = 2L,
  api_version = "v2",
  parser = "simdjson"
)
}
\arguments{
\item{file_path}{A string specifying the path to the JSONL file.}

\item{num_threads}{An integer specifying the number of threads to use. If \code{NULL}, it detects the number of cores automatically.}

\item{n_max}{Maximum number of lines to read from the file. Defaults to \code{Inf} (no limit).}

\item{skip_empty_rows}{Logical, whether to skip empty rows when reading the file. Defaults to \code{TRUE}.}

\item{query}{A character vector specifying the JSONPath to extract specific data. If \code{NULL}, it defaults based on the API version.}

\item{query_error_ok}{Logical, whether to ignore errors during query processing. Defaults to \code{TRUE}.}

\item{on_query_error}{A function or \code{NULL}, specifying the handler for query errors. Defaults to \code{NULL}.}

\item{parse_error_ok}{Logical, whether to ignore errors during parsing. Defaults to \code{TRUE}.}

\item{on_parse_error}{A function or \code{NULL}, specifying the handler for parsing errors. Defaults to \code{NULL}.}

\item{empty_array}{A value or \code{NULL}, specifying what to return for empty arrays. Defaults to \code{NULL}.}

\item{empty_object}{A value or \code{NULL}, specifying what to return for empty objects. Defaults to \code{NULL}.}

\item{max_simplify_lvl}{An integer specifying the maximum simplification level for the parsed JSON. Defaults to \code{2L}.}

\item{api_version}{A string specifying the API version, either \code{"v1"} or \code{"v2"}. Defaults to \code{"v2"}.}

\item{parser}{A character indicating which json parser should be used. Default "simdjson"; otherwise "yyjson".}
}
\value{
A data.table containing user information from the JSONL file.
The user ID column is renamed based on the API version, and duplicate rows are removed.
}
\description{
This function loads user data from a JSONL (JSON Lines) file and extracts relevant fields using the RcppSimdJson package.
It handles both v1 and v2 of the API by adjusting the query accordingly and flattens the resulting data.
}
