% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_batches.R
\name{load_users_jsonl_multi}
\alias{load_users_jsonl_multi}
\title{Load User Data from Multiple JSON Files}
\usage{
load_users_jsonl_multi(data_dir, ...)
}
\arguments{
\item{data_dir}{Directory containing the JSON files.}

\item{...}{Additional arguments to pass to \code{\link{load_tweets_jsonl}} function.}
}
\value{
A data.table containing tweets from all JSON files.
}
\description{
Load user data from multiple JSON files located in the specified directory.
}
