% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_line_tweets.R
\name{preprocess_line_tweets}
\alias{preprocess_line_tweets}
\title{Preprocess Line Tweets}
\usage{
preprocess_line_tweets(
  tweets,
  tweets_cols = c("possibly_sensitive", "lang", "text", "public_metrics_retweet_count",
    "public_metrics_reply_count", "public_metrics_like_count",
    "public_metrics_quote_count")
)
}
\arguments{
\item{tweets}{A data.table containing tweet data.}

\item{tweets_cols}{A character vector specifying the columns to keep in the processed data. Default columns include possibly_sensitive, lang, text, and various public metrics related to retweets, replies, likes, and quotes.}
}
\value{
A list containing data.tables for tweets, referenced tweets, URLs, mentions, and hashtags.
}
\description{
This function preprocesses tweet data, extracting relevant information such as entities, URLs, mentions, and hashtags, and organizes them into separate data.tables.
}
\details{
This function preprocesses tweet data, ensuring that required columns such as entities, tweet_id, created_at, author_id, conversation_id, and referenced_tweets are present. It then constructs the main data.table containing all tweets and their metadata. Public metrics are handled separately, and datetime columns are reformatted. Referenced tweets, entities, URLs, mentions, and hashtags are extracted and organized into separate data.tables for easy analysis.
}
