% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_line_tweets.R
\name{preprocess_line_users}
\alias{preprocess_line_users}
\title{Preprocess Line Users Data}
\usage{
preprocess_line_users(users, api_version = "v2")
}
\arguments{
\item{users}{A data.table containing user data from the Twitter API.}

\item{api_version}{A character string specifying the API version to handle, either \code{"v1"} or \code{"v2"}. Defaults to \code{"v2"}.}
}
\value{
A data.table with the preprocessed user data, including public metrics and URLs if applicable.
The \code{created_at} field is reformatted to a datetime object, and any extracted entities and URLs
are merged back into the main dataset.
}
\description{
This function preprocesses user data from the Twitter API, handling both v1 and v2 versions.
It ensures the necessary columns are present, processes public metrics, extracts URLs,
and reformats the \code{created_at} timestamp.
}
\details{
\itemize{
\item If the API version is \code{"v1"}, the function renames the \code{screen_name} column to \code{username}.
\item For API version \code{"v2"}, it unpacks the \code{public_metrics} and extracts URLs from the \code{entities} field.
\item The function checks for required columns (\code{username}, \code{user_id}) and stops if they are missing.
\item The \code{created_at} timestamp is reformatted based on the API version.
}
}
