% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_line_tweets.R
\name{rbind_preprocessed_dts}
\alias{rbind_preprocessed_dts}
\title{Row-bind Preprocessed Data Tables}
\usage{
rbind_preprocessed_dts(dt_list)
}
\arguments{
\item{dt_list}{A nested list containing preprocessed data.tables.}
}
\value{
A list containing row-binded data.tables with deduplication by
tweet_id for each type of data.
}
\description{
Combine preprocessed data.tables stored in a nested list into a list of row-binded data.tables,
deduplicating them by tweet_id.
}
