% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_line_tweets.R
\name{extract_URL_dt}
\alias{extract_URL_dt}
\title{Extract URLs Data.Table from Entities Data.Table}
\usage{
extract_URL_dt(entities, api_version = "v2")
}
\arguments{
\item{entities}{A data.table containing entities information extracted from Twitter data, typically obtained using the extract_entities_dt function.}

\item{api_version}{A character vector specifying the API version. Default is v2.}
}
\value{
A data.table containing URLs shared in the tweets, along with related information such as tweet_id, row_id, URL details, and domain names.
}
\description{
This function extracts URLs shared in tweets from the entities data.table and organizes them into a separate data.table.
}
\details{
This function evaluates if there are tweets with several URLs, resulting in a urls.url column. It then subsets the data into tweets
containing one URL and tweets with several URLs. Column names starting with "url." are harmonized, and the prefix "urls." is removed.
Empty observations are dropped, and a new column clear_url is created using data.table syntax. Domain names are extracted from the expanded_url column and stored in the domain column.
The resulting data.table is indexed by tweet_id, domain, and expanded_url for efficient querying.
}
