% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_line_tweets.R
\name{preprocess_line_tweets}
\alias{preprocess_line_tweets}
\title{Preprocess Line Tweets}
\usage{
preprocess_line_tweets(
  tweets,
  tweets_cols = c("possibly_sensitive", "lang", "text", "public_metrics_retweet_count",
    "public_metrics_reply_count", "public_metrics_like_count",
    "public_metrics_quote_count"),
  api_version = "v2"
)
}
\arguments{
\item{tweets}{A data.table of tweet data.}

\item{tweets_cols}{A character vector of column names to retain in the tweet data.
Default columns are provided, including possibly_sensitive, lang, and several public metrics.}

\item{api_version}{A character string specifying the API version. Must be either \code{"v1"} or \code{"v2"}.}
}
\value{
A list containing processed tweet data (tweets), referenced tweets (referenced),
URLs (urls), mentions (mentions), and hashtags (hashtags).
}
\description{
This function preprocesses tweet data, handling both v1 and v2 of the API.
It extracts and reshapes various tweet entities such as user information,
public metrics, URLs, mentions, and hashtags. It also adjusts the date format
and constructs the referenced tweets.
}
